+++ packages/google_maps_flutter/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapController.java
@@ -21,6 +21,10 @@ import android.graphics.Point;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
+import androidx.annotation.NonNull;
+import androidx.lifecycle.DefaultLifecycleObserver;
+import androidx.lifecycle.Lifecycle;
+import androidx.lifecycle.LifecycleOwner;
 import com.google.android.gms.maps.CameraUpdate;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.GoogleMapOptions;
@@ -34,9 +38,10 @@ import com.google.android.gms.maps.model.MapStyleOptions;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.Polygon;
 import com.google.android.gms.maps.model.Polyline;
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
+import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
-import io.flutter.plugin.common.PluginRegistry;
 import io.flutter.plugin.platform.PlatformView;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -48,6 +53,8 @@ import java.util.concurrent.atomic.AtomicInteger;
 /** Controller of a single GoogleMaps MapView instance. */
 final class GoogleMapController
     implements Application.ActivityLifecycleCallbacks,
+        DefaultLifecycleObserver,
+        ActivityPluginBinding.OnSaveInstanceStateListener,
         GoogleMap.OnCameraIdleListener,
         GoogleMap.OnCameraMoveListener,
         GoogleMap.OnCameraMoveStartedListener,
@@ -68,7 +75,6 @@ final class GoogleMapController
   private final int id;
   private final AtomicInteger activityState;
   private final MethodChannel methodChannel;
-  private final PluginRegistry.Registrar registrar;
   private final MapView mapView;
   private GoogleMap googleMap;
   private boolean trackCameraPosition = false;
@@ -81,7 +87,9 @@ final class GoogleMapController
   private final float density;
   private MethodChannel.Result mapReadyResult;
   private final int registrarActivityHashCode;
+  private final Lifecycle lifecycle;
   private final Context context;
+  private final Application application;
   private final MarkersController markersController;
   private final PolygonsController polygonsController;
   private final PolylinesController polylinesController;
@@ -95,18 +103,21 @@ final class GoogleMapController
       int id,
       Context context,
       AtomicInteger activityState,
-      PluginRegistry.Registrar registrar,
+      BinaryMessenger binaryMessenger,
+      Application application,
+      Lifecycle lifecycle,
+      int registrarActivityHashCode,
       GoogleMapOptions options) {
     this.id = id;
     this.context = context;
     this.activityState = activityState;
-    this.registrar = registrar;
     this.mapView = new MapView(context, options);
     this.density = context.getResources().getDisplayMetrics().density;
-    methodChannel =
-        new MethodChannel(registrar.messenger(), "plugins.flutter.io/google_maps_" + id);
+    methodChannel = new MethodChannel(binaryMessenger, "plugins.flutter.io/google_maps_" + id);
     methodChannel.setMethodCallHandler(this);
-    this.registrarActivityHashCode = registrar.activity().hashCode();
+    this.application = application;
+    this.lifecycle = lifecycle;
+    this.registrarActivityHashCode = registrarActivityHashCode;
     this.markersController = new MarkersController(methodChannel);
     this.polygonsController = new PolygonsController(methodChannel);
     this.polylinesController = new PolylinesController(methodChannel, density);
@@ -152,7 +163,11 @@ final class GoogleMapController
         throw new IllegalArgumentException(
             "Cannot interpret " + activityState.get() + " as an activity state");
     }
-    registrar.activity().getApplication().registerActivityLifecycleCallbacks(this);
+    if (lifecycle != null) {
+      lifecycle.addObserver(this);
+    } else {
+      application.registerActivityLifecycleCallbacks(this);
+    }
     mapView.getMapAsync(this);
   }
 
@@ -368,6 +383,10 @@ final class GoogleMapController
           result.success(googleMap.isBuildingsEnabled());
           break;
         }
+      case "map#getZoomLevel":
+        {
+          result.success(googleMap.getCameraPosition().zoom);
+        }
       case "map#setStyle":
         {
           String mapStyle = (String) call.arguments;
@@ -472,7 +491,7 @@ final class GoogleMapController
     disposed = true;
     methodChannel.setMethodCallHandler(null);
     mapView.onDestroy();
-    registrar.activity().getApplication().unregisterActivityLifecycleCallbacks(this);
+    application.unregisterActivityLifecycleCallbacks(this);
   }
 
   // @Override
@@ -489,6 +508,7 @@ final class GoogleMapController
     // TODO(mklim): Remove this empty override once https://github.com/flutter/flutter/issues/40126 is fixed in stable.
   };
 
+  // Application.ActivityLifecycleCallbacks methods
   @Override
   public void onActivityCreated(Activity activity, Bundle savedInstanceState) {
     if (disposed || activity.hashCode() != registrarActivityHashCode) {
@@ -543,6 +563,72 @@ final class GoogleMapController
       return;
     }
     mapView.onDestroy();
+  }
+
+  // DefaultLifecycleObserver and OnSaveInstanceStateListener
+
+  @Override
+  public void onCreate(@NonNull LifecycleOwner owner) {
+    if (disposed) {
+      return;
+    }
+    mapView.onCreate(null);
+  }
+
+  @Override
+  public void onStart(@NonNull LifecycleOwner owner) {
+    if (disposed) {
+      return;
+    }
+    mapView.onStart();
+  }
+
+  @Override
+  public void onResume(@NonNull LifecycleOwner owner) {
+    if (disposed) {
+      return;
+    }
+    mapView.onResume();
+  }
+
+  @Override
+  public void onPause(@NonNull LifecycleOwner owner) {
+    if (disposed) {
+      return;
+    }
+    mapView.onResume();
+  }
+
+  @Override
+  public void onStop(@NonNull LifecycleOwner owner) {
+    if (disposed) {
+      return;
+    }
+    mapView.onStop();
+  }
+
+  @Override
+  public void onDestroy(@NonNull LifecycleOwner owner) {
+    if (disposed) {
+      return;
+    }
+    mapView.onDestroy();
+  }
+
+  @Override
+  public void onRestoreInstanceState(Bundle bundle) {
+    if (disposed) {
+      return;
+    }
+    mapView.onCreate(bundle);
+  }
+
+  @Override
+  public void onSaveInstanceState(Bundle bundle) {
+    if (disposed) {
+      return;
+    }
+    mapView.onSaveInstanceState(bundle);
   }
 
   // GoogleMapOptionsSink methods