+++ packages/google_maps_flutter/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapFactory.java
@@ -4,10 +4,11 @@
 
 package io.flutter.plugins.googlemaps;
 
-import static io.flutter.plugin.common.PluginRegistry.Registrar;
-
+import android.app.Application;
 import android.content.Context;
+import androidx.lifecycle.Lifecycle;
 import com.google.android.gms.maps.model.CameraPosition;
+import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.StandardMessageCodec;
 import io.flutter.plugin.platform.PlatformView;
 import io.flutter.plugin.platform.PlatformViewFactory;
@@ -17,12 +18,23 @@ import java.util.concurrent.atomic.AtomicInteger;
 public class GoogleMapFactory extends PlatformViewFactory {
 
   private final AtomicInteger mActivityState;
-  private final Registrar mPluginRegistrar;
+  private final BinaryMessenger binaryMessenger;
+  private final Application application;
+  private final int activityHashCode;
+  private final Lifecycle lifecycle;
 
-  GoogleMapFactory(AtomicInteger state, Registrar registrar) {
+  GoogleMapFactory(
+      AtomicInteger state,
+      BinaryMessenger binaryMessenger,
+      Application application,
+      Lifecycle lifecycle,
+      int activityHashCode) {
     super(StandardMessageCodec.INSTANCE);
     mActivityState = state;
-    mPluginRegistrar = registrar;
+    this.binaryMessenger = binaryMessenger;
+    this.application = application;
+    this.activityHashCode = activityHashCode;
+    this.lifecycle = lifecycle;
   }
 
   @SuppressWarnings("unchecked")
@@ -48,6 +60,7 @@ public class GoogleMapFactory extends PlatformViewFactory {
     if (params.containsKey("circlesToAdd")) {
       builder.setInitialCircles(params.get("circlesToAdd"));
     }
-    return builder.build(id, context, mActivityState, mPluginRegistrar);
+    return builder.build(
+        id, context, mActivityState, binaryMessenger, application, lifecycle, activityHashCode);
   }
 }