+++ packages/google_maps_flutter/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.5.22
+
+* Support Android v2 embedding.
+* Bump the min flutter version to `1.12.13+hotfix.5`.
+
 ## 0.5.21+17
 
 * Fix Swift example in README.md.